Renamed package so it better describes the content
