+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRhymer.java
@@ -1,5 +1,5 @@
 package edu.kis.vh.nursery;
-import edu.kis.vh.nursery.list.*;
+import edu.kis.vh.nursery.stacks.*;
 
 
 public class DefaultCountingOutRhymer {
