+++ src/main/java/edu/kis/vh/nursery/stacks/IntStack.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.stacks;
 
 public interface IntStack {
 		void push(int i);
